Add files via upload
This code shows a multitude of functions and properties that allow for simple data manipulation and querying. By first creating a structure of tables and then filling those tables with values we can request information or study patterns that will give us insight on how we can improve our overall quality, services, work methods, and decision-making.